requirements.txt
attrs==19.3.0
Django
django-extensions==2.2.9
djangorestframework==3.11.0
Faker==4.1.0
importlib-metadata==1.6.0
more-itertools==8.3.0
packaging==20.4
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.7
pytest==5.4.2
pytest-django==3.9.0
python-dateutil==2.8.1
six==1.14.0
text-unidecode==1.3
wcwidth==0.1.9
zipp==3.1.0
